Api完善

+++ src/main/java/com/xie/controller/api/CartController.java
@@ -2,6 +2,9 @@ package com.xie.controller.api;
 
 import com.xie.response.BaseResponse;
 import com.xie.service.CartService;
+import com.xie.service.ItemService;
+import com.xie.service.ItemSpecService;
+import com.xie.utils.MallConstants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -15,6 +18,12 @@ public class CartController extends BaseController {
 
     @Autowired
     private CartService cartService;
+
+    @Autowired
+    private ItemService itemService;
+
+    @Autowired
+    private ItemSpecService itemSpecService;
 
     @RequestMapping(value = "/clear", method = RequestMethod.PUT)
     @ResponseBody
@@ -33,12 +42,21 @@ public class CartController extends BaseController {
     @ResponseBody
     public BaseResponse update(@RequestHeader(value = "sessionId") String sessionId, @RequestParam("gid") int gid, @RequestParam("spec") int spec, @RequestParam("amount") int amount) {
         int uid = getUid(sessionId);
-        if (amount > 0) {
-            cartService.saveOrUpdate(uid, gid, spec, amount);
-        } else {
-            cartService.deleteByGidAndSpec(gid, spec);
+        if (itemService.getById(gid).getIs_online() == MallConstants.NO || itemSpecService.getById(spec).getIs_online() == MallConstants.NO) {
+            return BaseResponse.fail("商品已经下架");
         }
-        return BaseResponse.ok(cartService.getByUidWithItem(uid));
+        int result = -1;
+        if (amount > 0) {
+            result = cartService.saveOrUpdate(uid, gid, spec, amount);
+        } else {
+            result = cartService.deleteByGidAndSpec(gid, spec);
+        }
+        if (result > 0) {
+            return BaseResponse.ok(cartService.getByUidWithItem(uid));
+        } else {
+            return BaseResponse.ok(cartService.getByUidWithItem(uid));
+        }
+
     }
 
     @RequestMapping(value = "/insert", method = RequestMethod.POST)
